test(core): factor view mutation acid test harness
